Renamed mixin collision.js to polygon.js.
